chore: move DomainError to error folder
