+++ app/api/core/domain/error/DomainError.ts
